Add compress meta api
